requirements.txt
faiss-cpu
pandas
numpy
sentence-transformers
transformers
fastapi
uvicorn
openai
